robotic-advisor-app/backend/services/llm_service.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure the Gemini API
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

def analyze_video(video_url):
    """
    Analyze a video using Gemini API to identify tasks and processes.
    
    Args:
        video_url (str): URL of the video to analyze
        
    Returns:
        dict: Analysis results including tasks, processes, and recommendations
    """
    # This is a placeholder for the actual implementation
    # In a real implementation, you would:
    # 1. Call Gemini API with the video URL or content
    # 2. Process the response to extract relevant information
    
    # For now, return a mock response
    return {
        "tasks": [
            {"name": "Pick and place", "duration": 45, "complexity": "medium"},
            {"name": "Assembly", "duration": 120, "complexity": "high"},
            {"name": "Inspection", "duration": 30, "complexity": "low"}
        ],
        "processes": [
            {"name": "Material handling", "automation_potential": "high"},
            {"name": "Quality control", "automation_potential": "medium"}
        ],
        "environment": {
            "space_constraints": "medium",
            "safety_concerns": "low"
        }
    }

def generate_recommendations(analysis_results, robot_options):
    """
    Generate recommendations based on analysis results and available robot options.
    
    Args:
        analysis_results (dict): Results from video analysis
        robot_options (list): Available robot options from URDF parsing
        
    Returns:
        list: Recommended robot configurations with justifications
    """
    # This would call Gemini API to generate recommendations
    # For now, return mock recommendations
    return [
        {
            "robot_id": robot_options[0]["id"],
            "configuration": "Standard arm with gripper",
            "justification": "Well-suited for the pick and place tasks identified in the video",
            "estimated_efficiency": 0.85
        },
        {
            "robot_id": robot_options[1]["id"],
            "configuration": "Precision arm with vision system",
            "justification": "Optimal for the assembly and inspection tasks requiring high precision",
            "estimated_efficiency": 0.92
        }
    ]