robotic-advisor-app/backend/requirements.txt
Flask>=2.0
google-generativeai
python-dotenv
Flask-Cors